Initial Commit
Includes the USB Host Shield Libraries that work with this code.  The newest version of this library broke something in the code that makes the controller fail to pair with the USB dongle most of the time.  I don't know what is broken.  However, if you use this library it works.  I also included the Syren and Sabertooth libraries.  I don't know if the newer versions work, but these do.